move mask related processing into a new file under utils/masks_utils.py
